fix(search): cast hybrid_score to float8 for correct sqlc type
sqlc was inferring hybrid_score as int32 instead of float64.
Added explicit CAST(... AS float8) to ensure proper float type generation.

+++ internal/services/search/client.go
@@ -124,7 +124,7 @@ func (c *Client) SearchHybrid(ctx context.Context, query string, limit int32) ([
 			ID:                pgUUIDToString(r.ID),
 			RecipeName:        r.RecipeName,
 			Description:       r.Description.String,
-			Similarity:        float64(r.HybridScore),
+			Similarity:        r.HybridScore,
 			CuisineCategories: interfaceToStringSlice(r.CuisineCategories),
 			MealTypes:         interfaceToStringSlice(r.MealTypes),
 		}
